OI4-239 - Removed unused web proxy and refactored name. Removed obsolete test f… (#46)
* Removed unused web proxy and refactored name. Removed obsolete test folder in src

* Refactored class and adjusted tests

* Removed package.json from include (for now) and fixed import

* Fixed merge problems

* Added a break into a switch case which I forgot to add while refactoring the class (#47)


* Removed unused web proxy and refactored name. Removed obsolete test folder in src

* Refactored class and adjusted tests

* Removed package.json from include (for now) and fixed import

* Fixed merge problems
